// src/app/models/customer/dao/customer-type-three.ts
export interface CustomerTypeThreeUpdate {

    firstName: string;

    secondName: string;

    lastName: string;

    secondLastName: string;
    
    documentTypeIdMaster: number;

    document: string;

    address: string;

    departmentIdDane: string;

    cityIdDane: string;

    email: string;

    phoneNumber: string;
}

export interface CustomerTypeThree {
}